refactor(web): move forms to forms folder
