+++ src/components/modules/ConvoMenu.tsx
@@ -6,7 +6,7 @@ import axios from 'axios';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useSession } from 'next-auth/react';
-import { ConvoForm } from './ConvoForm';
+import { ConvoForm } from '../forms/ConvoForm';
 
 const ConvoMenu = ({
   id,
